added exception handling

+++ book-rating-service/src/main/java/com/edc/pps/rating/service/RatingService.java
@@ -3,6 +3,7 @@ package com.edc.pps.rating.service;
 import com.edc.pps.rating.dto.RatingMapper;
 import com.edc.pps.rating.dto.RatingRequest;
 import com.edc.pps.rating.dto.RatingResponse;
+import com.edc.pps.rating.exception.RatingNotFoundException;
 import com.edc.pps.rating.model.Rating;
 import com.edc.pps.rating.repository.RatingRepository;
 import org.slf4j.Logger;
@@ -11,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
 @Service
@@ -22,7 +24,6 @@ public class RatingService {
     private final RatingRepository ratingRepository;
 
 
-
     @Autowired
     public RatingService(RatingRepository ratingRepository, RatingMapper ratingMapper) {
         this.ratingRepository = ratingRepository;
@@ -32,23 +33,23 @@ public class RatingService {
     public RatingResponse saveOrUpdate(RatingRequest request) {
         log.info("saved rating to db: {}", request);
         //get all ratings from db
-       List<Rating> ratings = ratingRepository.findAll();
-       //check if the user already rated the book
-      long result = ratings.stream().filter(entry -> entry.getBookId() == request.getBookId() && entry.getUserId() == request.getUserId()).count();
+        List<Rating> ratings = ratingRepository.findAll();
+        //check if the user already rated the book
+        long result = ratings.stream().filter(entry -> entry.getBookId() == request.getBookId() && entry.getUserId() == request.getUserId()).count();
 
         //if the user never rated the book it will be added as a new entry in db
-      if(result == 0) {
-          Rating rating = ratingMapper.toEntity(request);
-          ratingRepository.save(rating);
-          return ratingMapper.toDto(rating);
-      }
-      //if the user already rated the book it will be updated
-        else{
-          Rating foundRating =  ratingRepository.findByBookIdAndUserId(request.getBookId(),request.getUserId());
-          foundRating.setRatingValue(request.getRatingValue());
-          ratingRepository.save(foundRating);
-        return ratingMapper.toDto(foundRating);
-      }
+        if (result == 0) {
+            Rating rating = ratingMapper.toEntity(request);
+            ratingRepository.save(rating);
+            return ratingMapper.toDto(rating);
+        }
+        //if the user already rated the book it will be updated
+        else {
+            Rating foundRating = ratingRepository.findByBookIdAndUserId(request.getBookId(), request.getUserId());
+            foundRating.setRatingValue(request.getRatingValue());
+            ratingRepository.save(foundRating);
+            return ratingMapper.toDto(foundRating);
+        }
     }
 
 
@@ -59,18 +60,25 @@ public class RatingService {
     }
 
     public void delete(Long id) {
-        log.debug("deleting rating with id: {}", id);
-        ratingRepository.deleteById(id);
+        //1. rating with id was not found
+        Optional<Rating> rating = ratingRepository.findByRatingId(id);
+        if (rating.isEmpty()) {
+            throw new RatingNotFoundException("no rating with id: " + id);
+        } else {
+            log.debug("deleting rating with id: {}", id);
+            ratingRepository.deleteById(id);
+        }
+
     }
 
     //method to get all ratings for a certain book to calculate avergae
 
-    public List<RatingResponse> getAllRatingsForBook(long bookId){
-        //find all books where book_id = ?
+    public List<RatingResponse> getAllRatingsForBook(long bookId) {
         log.debug("getting all rating for bookId: {}", bookId);
-    List<Rating> ratings = ratingRepository.findByBookId(bookId);
-    return ratingMapper.toDto(ratings) ;
+        List<Rating> ratings = ratingRepository.findByBookId(bookId);
+        return ratingMapper.toDto(ratings);
     }
+
 
 
 }
